requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-annotate requirements.in
#
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
automat==20.2.0
cachecontrol==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
channels-redis==3.2.0
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django==3.2.3
djangorestframework==3.12.4
firebase-admin==5.0.0
google-api-core[grpc]==1.30.0
google-api-python-client==2.8.0
google-auth-httplib2==0.1.0
google-auth==1.31.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
hiredis==2.0.0
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
psycopg2-binary==2.8.6
psycopg2==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyopenssl==20.0.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
twisted[tls]==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
zope.interface==5.4.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
